complete homework 1.2
