+++ src/AddToDoForm.js
@@ -1,0 +1,13 @@
+import React from "react";
+
+function AddToDoForm(){
+    return(
+        <form>
+            <label htmlFor='todoTitle'>title</label>
+            <input id='todoTitle'></input>
+            <button type='submit'>Add</button>
+        </form>
+    )
+}
+
+export default AddToDoForm
