+++ src/App.js
@@ -1,37 +1,22 @@
 import React from 'react';
+import ToDoList from './TodoList.js';
+import AddToDoForm from './AddToDoForm.js';
 
-const todoList = [
-  {
-    id: 1,
-    title: "wash face"
-  },
-  {
-    id: 2,
-    title: "make coffee"
-  },
-  {
-    id: 3,
-    title: "physical therapy"
-  }
-]
 
 function App() {
-
-
   return (
     <div>
 
       <h1>To Do List</h1>
-
-      <ul>
-        {todoList.map(function (item) {
-          return <li key={item.objectID}>{item.title}</li>;
-        })}
-      </ul>
-
+      <ToDoList />
+      <AddToDoForm />
 
     </div>
   );
-}
+};
+
+
+
 
 export default App;
+
